modify: 렌더링 최적화 - 내부 컴포넌트 코드를 components 디렉토리로 분리 - Purchase 컨테이너 추가 작업 필요 : 수량 조정 시 전체 리렌더링링
